Kernel: Refactor CP15 R/W
All read/writes to coprocessor 15 now go through the cp15.h interface.
